Refactor map to separate the heatmap generation from the rest of the map
The class are tightly coupled but I think that's unavoidable
